Fix path_esc for linux builds
As near as I can tell, path_esc is only called in one place and it's always called with a string argument. Yet, it appears to be using the argument as an array. This is causing failures when attempting to run jar on Debian; it complains that 'map' doesn't exist on the String.